requirements.txt
matplotlib
pytorch_toolbelt
tqdm
fvcore
timm
prettytable
einops
